# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Go Boilerplate",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0.4"
    },
    "paths": {
        "/api/v1/comet": {
            "get": {
                "description": "Fetch CometScraper",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CometScrapers"
                ],
                "summary": "Fetch CometScraper",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "description": "Create CometScraper",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CometScrapers"
                ],
                "summary": "Create CometScraper",
                "parameters": [
                    {
                        "description": "CometScraper to create",
                        "name": "car",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateCometScraperReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/comet/{id}": {
            "get": {
                "description": "Get CometScraper",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CometScrapers"
                ],
                "summary": "Get CometScraper",
                "parameters": [
                    {
                        "type": "string",
                        "description": "car id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Delete CometScraper",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CometScrapers"
                ],
                "summary": "Delete CometScraper",
                "parameters": [
                    {
                        "type": "string",
                        "description": "car id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "request.CreateCometScraperReq": {
            "type": "object",
            "properties": {
                "email": { "type": "string"},
                "password": { "type": "string"}
            }
        }
    }
}